move jest-ish functions to own module
